Encapsulate the certificate sending behavior
Previously we needed to call send_certificate and deliver certificate
explicitly, but sending certificate is a wrapper on top of kit's save
method. We use send_certificate when we need to mark the kit as used
and also deliver the certificate. So let's encapsulate these behavior
to the send_certificate public api