Change certificate content creation process
Certificate content creation page is expecting user to edit
the content using visual tools and once they are ready then
submit that for approval. But the problem is most of user
tries to submit it without adding necessary content.

So let's change this process, now if user tries to submit a
certificate without required content then it would not allow
it unless they use the visual tool to add/edit their content